split infrastructure into directories
