+++ TagsCloudContainer/App/CloudGenerator/CloudGenerator.cs
@@ -2,7 +2,7 @@
 using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
-using TagsCloudContainer.Infrastructure;
+using TagsCloudContainer.Infrastructure.CloudGenerator;
 
 namespace TagsCloudContainer.App.CloudGenerator
 {
