+++ TagsCloudContainer/App/FrequencyDictionaryGenerator/SimpleTextParser.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using TagsCloudContainer.Infrastructure;
+using TagsCloudContainer.Infrastructure.DictionaryGenerator;
 
 namespace TagsCloudContainer.App.FrequencyDictionaryGenerator
 {
