+++ TagsCloudContainer/Infrastructure/CloudVisualizer/IImageGenerator.cs
@@ -3,7 +3,7 @@ using System.Drawing;
 using TagsCloudContainer.App.CloudGenerator;
 using TagsCloudContainer.App.Settings;
 
-namespace TagsCloudContainer.Infrastructure
+namespace TagsCloudContainer.Infrastructure.CloudVisualizer
 {
     internal interface IImageGenerator
     {
