+++ TagsCloudContainer/Infrastructure/DictionaryGenerator/IFrequencyDictionaryGenerator.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace TagsCloudContainer.Infrastructure
+namespace TagsCloudContainer.Infrastructure.DictionaryGenerator
 {
     internal interface IFrequencyDictionaryGenerator
     {
