+++ TagsCloudContainer/Infrastructure/DictionaryGenerator/ITextParser.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace TagsCloudContainer.Infrastructure
+namespace TagsCloudContainer.Infrastructure.DictionaryGenerator
 {
     internal interface ITextParser
     {
