+++ TagsCloudContainer/Infrastructure/FileReader/IFileReader.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace TagsCloudContainer.Infrastructure
+namespace TagsCloudContainer.Infrastructure.FileReader
 {
     internal interface IFileReader
     {
